change process ID type
`len(p)` in Go returns `int` (odd that it doesn't use an unsigned
integer type), so switching to use `int` for process IDs will be less of
an impedance mismatch with how I plan to issue process IDs.

I'll continue to use `uint32` on the ProtoBuf side though. Ideally,
I'd add a `(ProcessManager).Len` method that can be used to ensure new
processes won't have an ID outside the bounds of `uint32` but I don't
think that's really necessary for this project. Also, that would require
spawning a ton of processes.