use stretchr/testify
This indeed makes test functions easier/shorter to write, which is nice.

Unfortunately Go's lockfile format has some quirks that make it not play
nice with Nix, so we have to remember to update a hash all the time to
use buildGoModule. See <https://www.tweag.io/blog/2021-03-04-gomod2nix/>